+++ lib/rubocop/cop/rspec_cops.rb
@@ -4,7 +4,11 @@ require_relative 'rspec/capybara/feature_methods'
 require_relative 'rspec/factory_bot/dynamic_attribute_defined_statically'
 require_relative 'rspec/factory_bot/static_attribute_defined_dynamically'
 
-require_relative 'rspec/rails/http_status'
+begin
+  require_relative 'rspec/rails/http_status'
+rescue LoadError # rubocop:disable Lint/HandleExceptions
+  # Rails/HttpStatus cannot be loaded if rack/utils is unavailable.
+end
 
 require_relative 'rspec/align_left_let_brace'
 require_relative 'rspec/align_right_let_brace'