+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * The `Rails/HttpStatus` cop is unavailable if the `rack` gem cannot be loaded. ([@bquorning][])
 * Fix `Rails/HttpStatus` not working with custom HTTP status codes. ([@bquorning][])
 * Fix `FactoryBot/StaticAttributeDefinedDynamically` to handle empty block. ([@abrom][])
+* Fix false positive in `FactoryBot/DynamicAttributeDefinedStatically` when a before/after callback has a symbol proc argument. ([@abrom][])
 
 ## 1.23.0 (2018-02-23)